+++ NEWS.md
@@ -1,5 +1,5 @@
 # Version 1.0.9000.9098
-* Renamed `stopList` slot in `StoppingList` class.
+* Renamed `stopList` slot in `StoppingList` and `StoppingAll` classes.
 * Added `unique` flag to `assert_probabilities` checkmate custom functions.
 * In `StoppingTargetBiomarker` class, changed `scale` character slot to
   `is_relative` flag.